Add user details page and move edit page
