+++ frontend/app/(main)/users/[id]/page.tsx
@@ -1,150 +1,82 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { useParams, useRouter } from 'next/navigation';
+import { useParams } from 'next/navigation';
+import Link from 'next/link';
 import AuthGuard from '../../../../components/AuthGuard';
 import Spinner from '../../../../components/Spinner';
 import api from '../../../../lib/api';
-
-interface ApiRole {
-  id: number;
-  name: string;
-}
+import { useAuth } from '../../../../context/AuthContext';
 
 interface ApiUser {
   id: number;
   firstName: string;
   lastName: string;
-  username: string;
-  role: { id: number; name: string };
+  email: string;
+  role: { name: string };
+  createdAt: string;
 }
 
-export default function UserEditPage() {
+export default function UserDetailsPage() {
   const params = useParams<{ id: string }>();
-  const router = useRouter();
-
-  const [roles, setRoles] = useState<ApiRole[]>([]);
+  const { user } = useAuth();
+  const [data, setData] = useState<ApiUser | null>(null);
   const [loading, setLoading] = useState(true);
-  const [saving, setSaving] = useState(false);
   const [error, setError] = useState('');
 
-  const [firstName, setFirstName] = useState('');
-  const [lastName, setLastName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [roleId, setRoleId] = useState<number | ''>('');
-
-  const fetchData = () => {
+  useEffect(() => {
     setLoading(true);
-    Promise.all([
-      api.get<ApiUser>(`/users/${params.id}`),
-      api.get<ApiRole[]>(`/roles`),
-    ])
-      .then(([userRes, rolesRes]) => {
-        const user = userRes.data;
-        setFirstName(user.firstName);
-        setLastName(user.lastName);
-        setEmail(user.username);
-        setRoleId(user.role.id);
-        setRoles(rolesRes.data);
+    api
+      .get<ApiUser>(`/users/${params.id}`)
+      .then(res => {
+        setData(res.data);
+        setError('');
       })
-      .catch(() => setError('Failed to load data'))
+      .catch(() => setError('Failed to load user'))
       .finally(() => setLoading(false));
-  };
+  }, [params.id]);
 
-  useEffect(fetchData, [params.id]);
-
-  const onSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setSaving(true);
-    setError('');
-    try {
-      const body: any = {
-        firstName,
-        lastName,
-        username: email,
-        roleId: Number(roleId),
-      };
-      if (password) body.password = password;
-      await api.put(`/users/${params.id}`, body);
-      router.push('/users');
-    } catch (err) {
-      setError('Failed to save user');
-    } finally {
-      setSaving(false);
-    }
-  };
+  const permissions = user?.permissions?.map((p: any) => p.code) || [];
+  const canEdit = permissions.includes('users:edit');
 
   return (
     <AuthGuard>
       {loading ? (
         <Spinner />
-      ) : (
-        <form onSubmit={onSubmit} className="space-y-4 max-w-md">
-          {error && <p className="text-red-500">{error}</p>}
-          <div>
-            <label className="block mb-1">First Name</label>
-            <input
-              type="text"
-              value={firstName}
-              onChange={e => setFirstName(e.target.value)}
-              className="w-full p-2 bg-[#1E1E1E] rounded"
-            />
+      ) : error ? (
+        <p className="text-red-500">{error}</p>
+      ) : data ? (
+        <div className="space-y-4">
+          <div className="bg-[#1E1E1E] p-4 rounded">
+            <h1 className="text-xl font-bold mb-4">
+              {data.firstName} {data.lastName}
+            </h1>
+            <table className="w-full text-sm">
+              <tbody>
+                <tr>
+                  <td className="p-2 font-semibold">Email</td>
+                  <td className="p-2">{data.email}</td>
+                </tr>
+                <tr>
+                  <td className="p-2 font-semibold">Role</td>
+                  <td className="p-2">{data.role.name}</td>
+                </tr>
+                <tr>
+                  <td className="p-2 font-semibold">Created At</td>
+                  <td className="p-2">
+                    {data.createdAt ? new Date(data.createdAt).toLocaleString() : ''}
+                  </td>
+                </tr>
+              </tbody>
+            </table>
           </div>
-          <div>
-            <label className="block mb-1">Last Name</label>
-            <input
-              type="text"
-              value={lastName}
-              onChange={e => setLastName(e.target.value)}
-              className="w-full p-2 bg-[#1E1E1E] rounded"
-            />
-          </div>
-          <div>
-            <label className="block mb-1">Email</label>
-            <input
-              type="email"
-              value={email}
-              onChange={e => setEmail(e.target.value)}
-              className="w-full p-2 bg-[#1E1E1E] rounded"
-            />
-          </div>
-          <div>
-            <label className="block mb-1">Password</label>
-            <input
-              type="password"
-              value={password}
-              onChange={e => setPassword(e.target.value)}
-              placeholder="Leave blank to keep current"
-              className="w-full p-2 bg-[#1E1E1E] rounded"
-            />
-          </div>
-          <div>
-            <label className="block mb-1">Role</label>
-            <select
-              value={roleId}
-              onChange={e => setRoleId(Number(e.target.value))}
-              className="w-full p-2 bg-[#1E1E1E] rounded"
-            >
-              <option value="" disabled>
-                Select Role
-              </option>
-              {roles.map(r => (
-                <option key={r.id} value={r.id}>
-                  {r.name}
-                </option>
-              ))}
-            </select>
-          </div>
-          <button
-            type="submit"
-            disabled={saving}
-            className="px-4 py-2 bg-accent text-black rounded disabled:opacity-50"
-          >
-            {saving ? 'Saving...' : 'Save Changes'}
-          </button>
-        </form>
-      )}
+          {canEdit && (
+            <Link href={`/users/${data.id}/edit`} className="px-4 py-2 bg-accent text-black rounded">
+              Edit
+            </Link>
+          )}
+        </div>
+      ) : null}
     </AuthGuard>
   );
 }
